Create a file to indicate reconstruction done
Earlier, to indicate a timestamp had been successfully reconstructed, I
was appending "done" to the directory name. To begin tarring the
directory, I then had to rename it back to the original directory name
(without the "done" suffix). The issue with this was that if the tarring
process got interrupted, I would have no way of knowing whether the
reconstruction was successful or not because it no longer had the "done"
suffix.

To get around this issue, I decided it would be better to create an
extra file inside the time directory to use as a marker. Then, when
tarring, I can just exclude the file while creating the tar, and just
leave it lying around so that it is always possible to know that the
reconstruction was successful. This commit updates all the Task objects
to use this method.