Increase independence of Tasks and checks
Prior to this commit, I felt that the code was a little too tightly
coupled when I thought about it; it was hard to see what checks /
conditions were triggering which actions / tasks. Through this commit, I
hope to have reduced some of this apparent coupling by moving some
things around and beginning the process of simplifying the Task
structure.

Tasks no longer have a list of "next_tasks". Instead, each task is
created based on some independent check / condition being satisfied. I
believe this will benefit in at least two ways. Firstly, it will allow
for the system to be more robust, especially when interrupted, as each
condition can be checked independently and incomplete tasks can be
re-attempted when the system is restarted. Secondly, I believe it will
allow the system to be extended to handle other conditions and interface
with other simulation frameworks in the long run. The simplification
process will continue in the future by further simplifying the Task
structure - I plan on possibly eliminating the specialized Task types
altogether in favor of passing in specialized behavior when constructing
the tasks.