Add optional completion_check method to Task
This commit allows Task to accept an optional completion_check function.
This function, when passed in, performs an additional check to figure
out if the Task has already been completed successfully or not. This
will allow the Task to skip over the entire run logic if it is already
completed.