Allow arbitrary priorities in PriorityList
This commit updates the PriorityList so that it is not constrained to
running on a finite set of priorities specified at creation time. I'm
not sure how the performance will change because of this but I believe,
for this application, it shouldn't make that much of a difference. But
it does simplify the code quite a bit and allows for extending the code
so that Tasks can now be created with arbitrary priorities without
needing to change anything else. Also, PriorityList objects can now be
created without having to specify the number of priorities allowed,
which is one less thing to worry about. Finally, the tests have been
simplified a lot as well!