Don't delete last available split timestep ever
There was a bug in when the split timesteps were being deleted. When
starting up a simulation, if there was only one split time and it was
also already reconstructed successfully, simon would decide to delete
the split timestep. But if the simulation had not started by this time,
it would cause the simulation to crash.

This commit forces simon to keep the last timestep around (which is also
good in case we want to restart, I guess).