Move processed file checks to get_new_tasks()
When OFListener is looking for more files, its method generate lists
containing only the files that have not been processed. For the next
step (compressing tarred files), I needed access to all the tarred files
currently available, not just the ones that hadn't yet been processed.
This commit makes that possible by moving the conditions that check
whether or not a file has been processed to where the actions are being
created instead of at the file list generation steps. This should make
the next part be easier to develop.