Update slurm compress submit command test
This commit updates the test that checks the command slurm runs to
submit the compress command. It is necessary that sbatch is run from the
case directory. So this commit ensures that the command being run begins
by cd-ing into the case directory and then resetting the working
directory. I don't think that last part is necessary, but I've added it
in just in case.