Restructure OFListener tests
After having split up the tests for OFListener, they definitely are much
easier to read. This commit groups tests together by idea and renames
the files and some of the tests to better reflect what they are doing.
Hopefully this will make it easier to find the right place to add tests
in the future.