Split up OFListener class
This commit splits up the OFListener class into an OFListener class and
an OFFileState class. The OFFileState class is responsible for analyzing
what files are present in the OpenFOAM simulation directory while the
OFListener class is responsible for figuring out what to do with this
given state of all the files. This change was done to help simplify the
OFListener class which was beginning to get pretty complex.