Create OFListener instance as a fixture
In each of the tests for the OFListener class, the OFListener object was
being created inside the test body. This was getting annoying because
every time I changed the parameters needed to create an instance of it,
I had to also change it in every single test. This commit just moves
that creation logic out into a fixture so that it's in one place now
(except for the tests that look at initialization).