Require that keep_every divides compress_every
In OFListener, the parameter keep_every defines how often files are
being reconstructed while the parameter compress_every defines how often
the files are being compressed. If compress_every is not a multiple of
keep_every, then there are files that are being reconstructed but not
compressed. I don't particularly want to throw reconstructed data away
at this point, so this commit ensures that that is not allowed. It also
makes some of the logic for compressing files easier, I think.