Ensure processing frequencies are constrained
In a previous commit, I ensured that OFListener's compress_every was a
multiple of keep_every. However, I failed to realize that these values
could also be set externally (which is something I was doing in some of
the testing code). This caused a lot of the tests to break while I was
working on the compression feature.

This commit introduces getters and setters for these properties so that
we can be sure that whenever they are updated, they are still sensible.