Allow updating processing frequencies together
With the new checks on keep_every needing to divide compress_every, if
we wanted to update both of these values to a valid combo, that would
require updating one and then the other. Let's say that we chose to
update keep_every first and then compress_every. If the new value of
keep_every did not divide into the old value of compress_every, we'd get
a ValueError. This is not desired since we are going to follow up with
an update to compress_every which would fix it.

This commit introduces a method that allows updating both values
simultaneously to get around this issue. The alternative would have been
pretty messy - trying to find GCDs and LCMs of these parameters and
doing the update in multiple stages which sounds awful.