Update OFListener fixture
Many of the tests that don't even rely on compress_every started failing
because as they were varying the keep_every parameter, they would run
into a test case where keep_every did not divide compress_every.

This commit modifies the compress_every value in the conftest listener
fixture to 3000 as it is large enough and is a multiple of all the test
cases used. The tests where this number really matters hopefully
override this value themselves.