Provide utilities to handle compressed filenames
This commit provides the OFFileState with two utility methods that allow
it to create a compressed file name from it's components and extract the
components from a given file name. This will hopefully mean that the
file name format is only defined in one place so that it's harder to
mess up.