Add config implementation
This commit introduces a very bare bones implementation of a config
module. This will allow most of simon's configuration options to be
provided through a config file instead of as command line arguments.
This should make it easier to use, especially on slurm clusters where
you can just put a config file in each simulation directory (similar to
an sfile, for example). I still need to work on allowing simon to read
values from the Config object and remove the corresponding options from
the command line interface.

I chose to implement it through the configparser module (which uses an
INI file) because I need to support python 3.9 on PACE. I think,
eventually it might be better to move to TOML or something else, but for
now, this should work just fine.